\name{3.1.plot.interactives}
\alias{3.1.plot.interactives}
\alias{getXY}
\alias{getLatLon}

\title{
  Interactive plot functions
}
\description{
  Recovering information from existing lattice plots.
}
\usage{

getXY(n = -1, ..., unit = "native", scale.correction = NULL)

getLatLon(..., map = NULL, object = trellis.last.object(),  
          scale.correction = function(x) {
                                temp <- XY2LatLon(map, x$x, x$y)
                                as.list(as.data.frame(temp))
                            })

}

\arguments{

  \item{n}{ If positive, the maximum number of points to locate. If negative 
   (default), unlimited. 
  } 

  \item{unit}{ The unit to use when reporting located points, by 
   default \code{"native"}.
  } 

  \item{scale.correction}{ The correction to apply if the plot has 
   locally scaled axes. See Note below.  
  } 

  \item{map, object}{ For \code{getLatLon} only. The plot layer as 
   generated \code{makeMapArg} and the plot object genereated by 
   \code{googleMap}. The \code{map} is strictly required as a reference when 
   converting plot points to associated latitude, longitude values, and 
   can be supplied directly as \code{map}, or recovered from the plot, 
   which can be supplied as \code{object}. If neither are supplied (as in 
   default use), the function attempts to recover \code{map} from the 
   last \code{lattice} plot via \code{trellis.last.object}.
   }

  \item{...}{ Additional arguments, passed on to related functions.

   These may be subject to revision, but are currently: 
   \code{\link{trellis.focus}} for panel selection (if working with multi-panel 
   plots) and \code{\link{lpoints}} to set point properties (if marking selected 
   points). For \code{getLatLon}, additional arguments are also passed to 
   \code{\link{XY2LatLon}} for x, y to latitude, longitude rescaling.
 
  } 

}


\details{
  \code{getXY} is an interactive function which returns 
  the locations of points on a plot selected using the 
  mouse (left click to select points; right click and 
  stop to end point collection; escape to abort without 
  returning any values). 

  It is a wrapper for the \code{\link{grid}} function 
   \code{\link{grid.locator}} that behaves more like 
   \code{\link{locator}}, the equivalent function intended 
   for use with \code{\link{plot}} outputs. 

  By default \code{getXY} selections are not automatically 
   marked. Adding common plot parameters to the function 
   call overrides this behaviour, e.g. to add red symbols 
   and lines.  
   
   \code{ans <- getXY(col = "red", pch = 4, type = "b")} 
   
  \code{getXY} also provides a mechanism to handle data plotted on 
   locally scaled axes. See Note below.

  \code{getLatLon} is wrapper for \code{getXY} for use 
   with \code{googleMap} outputs and other similarly 
   georeferenced plots. See Note below. 
     
}

\value{
  \code{getXY} returns the x and y coordinates of 
   the selected points on a plot as a list containing two 
   components, \code{x} and \code{y}.

  \code{getLatLon} returns the latitude and longitude 
   values of the selected points on a map as a list 
   containing two components, \code{lat} and \code{lon}.
  
}

\note{

  INTRO: local scaling

  Corrections are applied using \code{scale.correction} if it 
  is supplied. This should therefore be a function that can be 
  applied directly to a standard \code{getXY} output (a list of 
  \code{x} and \code{y} values) and that rescales these to give 
  their corrected values.

  \code{getLatLon} provides an example of the mechanism, and is 
  for use with georeferenced plots that have been locally scaled 
  using \code{\link{RgoogleMaps}} functions like 
  \code{\link{LatLon2XY}}. \code{getLatLon} uses \code{XY2LatLon} to 
  rescale \code{x} and \code{y} values and then \code{as...} functions 
  to convert the outputs of this step to a list format like that 
  generated by \code{locator}, \code{grid.locator} or \code{getXY}.    
}


\references{
  This function makes extensive use of code developed by others.

  lattice:
  Sarkar, Deepayan (2008) Lattice: Multivariate Data
  Visualization with R. Springer, New York. ISBN
  978-0-387-75968-5

  RgoogleMaps:
  Markus Loecher and Sense Networks (2011).
  RgoogleMaps: Overlays on Google map tiles in R. R
  package version 1.1.9.6.
  http://CRAN.R-project.org/package=RgoogleMaps
}

\author{
  Karl Ropkins
}

\seealso{
   For related functions in other packages, see

   \code{\link{lattice}}: \code{\link{grid.locator}}; 
   \code{\link{trellis.focus}} and \code{\link{lpoints}} 

   \code{\link{googleMap}}: \code{\link{LatLon2XY}}, etc.  
}

\examples{

##example 1
##to be confirmed

##getXY on xyplot

##example 2

##Move getLatLon code to examples




}


\keyword{ methods }
